test(global-header): playwright refactor
