+++ playwright/components/global-header/locators.ts
@@ -1,0 +1,4 @@
+// component preview locators
+export const GLOBAL_HEADER = "[data-component='global-header']";
+export const GLOBAL_HEADER_LOGO_WRAPPER =
+  "[data-element='global-header-logo-wrapper']";
